+++ core/modules-entry.ts
@@ -1,10 +1,11 @@
 
 import { VueStorefrontModule } from './modules'
 import { Wishlist } from './modules/wishlist'
-import { Mailer } from './modules/mailer'
-import { registerModules } from 'theme/modules'
+import { User } from './modules/user'
+import { registerModules } from 'src/modules'
 
 export const enabledModules: VueStorefrontModule[] = [
   Wishlist,
+  User,
   ...registerModules
 ]