+++ src/extensions/index.js
@@ -11,5 +11,4 @@ if (!Vue.prototype.$isServer) { // extensions that are not required in the SSR m
 extensionList.push(require('@vue-storefront/raw-output-example/index.js'))
 extensionList.push(require('@vue-storefront/extension-payment-backend-methods/index.js'))
 extensionList.push(require('@vue-storefront/extension-payment-cash-on-delivery/index.js'))
-extensionList.push(require('vsf-payment-stripe/index.js'))
 export default extensionList